requirements.txt
Flask==3.1.3
python-amazon-sp-api==2.1.8
python-dotenv==1.2.1
requests==2.32.5
gunicorn==23.0.0
